Move all CSS into a include file, reduce whitespace count on primary html
